This is a corrected Ex1 script.
The error was that the loop couldn't sum the sequences across all files. It was the way how I defined the variables inside the loop.
Therefore, I corrected the loop, instead of treating each file individually, I stored them into an array and applied the loop on the array. Seq1 variable has a string of values.
When I echo or sum Seq1, it gives the total number of sequences across all files not only the last value that the loop runs over as occured in the first script.
Although the observed output is exactly the same of the expected output, this script still gives an error.